Refactor storage classes into separate package
