+++ src/main/java/warehouse/storage/containers/StorageUnitContainer.java
@@ -1,4 +1,4 @@
-package warehouse;
+package warehouse.storage.containers;
 
 import inventory.Item;
